Reorganize backend components to backend folder, update all references
