+++ src/backend/components/categories/Categories.vue
@@ -75,7 +75,7 @@
 </template>
 
 <script>
-import collectionMixin from '../../mixins/collectionMixin';
+import collectionMixin from '@/mixins/collectionMixin';
 export default {
   components: {},
   mixins: [collectionMixin],
