+++ src/backend/components/combinators/Combinators.vue
@@ -37,7 +37,7 @@
 </template>
 
 <script>
-import collectionMixin from '../../mixins/collectionMixin'
+import collectionMixin from '@/mixins/collectionMixin'
 export default {
   data() {
     return {
