+++ src/backend/components/concept-maps/ConceptMaps.vue
@@ -100,7 +100,7 @@
 </template>
 
 <script>
-import collectionMixin from '../../mixins/collectionMixin';
+import collectionMixin from '@/mixins/collectionMixin';
 export default {
   mixins: [collectionMixin],
   data() {
