+++ src/backend/components/concepts/Concepts.vue
@@ -77,7 +77,7 @@
 </template>
 
 <script>
-import collectionMixin from '../../mixins/collectionMixin'
+import collectionMixin from '@/mixins/collectionMixin'
 export default {
   mixins: [collectionMixin],
   data() {
