+++ src/backend/components/index.js
@@ -1,0 +1,9 @@
+export { default as Combinators } from './combinators/Combinators'
+export { default as Categories } from './categories/Categories'
+export { default as Datasets } from './datasets/Datasets'
+export { default as Users } from './users/Users'
+export { default as DatasetFields } from './datasets/DatasetFields'
+export { default as TemporalCoverages } from './temporal-coverages/TemporalCoverages'
+export { default as ConceptMaps } from './concept-maps/ConceptMaps'
+export { default as ConceptTopics } from './concept-maps/ConceptTopics'
+export { default as Concepts } from './concepts/Concepts'
