+++ src/backend/components/temporal-coverages/TemporalCoverages.vue
@@ -62,7 +62,7 @@
 </template>
 
 <script>
-import collectionMixin from '../../mixins/collectionMixin'
+import collectionMixin from '@/mixins/collectionMixin'
 export default {
   mixins: [collectionMixin],
   data() {
