+++ src/backend/components/users/Users-index.vue
@@ -64,7 +64,7 @@
 
 <script>
 import { mapState, mapActions } from 'vuex'
-import collectionMixin from '../../mixins/collectionMixin'
+import collectionMixin from '@/mixins/collectionMixin'
 export default {
   mixins: [collectionMixin],
   data() {
