+++ src/backend/templates/TableViewLayout.vue
