+++ src/backend/views/Pages/ForgotPassword.vue
