--- src/models/index.js
@@ -1,9 +1,0 @@
-export { default as Combinators } from './Combinators'
-export { default as Categories } from './Categories'
-export { default as Datasets } from './Datasets'
-export { default as Users } from './Users'
-export { default as DatasetFields } from './DatasetFields'
-export { default as TemporalCoverages } from './TemporalCoverages'
-export { default as ConceptMaps } from './ConceptMaps'
-export { default as ConceptTopics } from './ConceptTopics'
-export { default as Concepts } from './Concepts'
